Quick and dirty implementation of constructor and get method
It uses non enumarable wrapped property to hold the passed object.

Get method traverses the wrapped object. It should support mapping, but there
are no tests for that yet.

Ultimately the object passed to constructor should be merged with data-object,
so it's properties could be accessed in a usual way in addition to get and set
methods.